+++ packages/chaingraph-nodes/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @badaitech/chaingraph-nodes
 
+## 0.1.20
+
+### Patch Changes
+
+- Bump version
+- Updated dependencies
+  - @badaitech/chaingraph-types@0.1.20
+
+## 0.1.19
+
+### Patch Changes
+
+- Bump package
+- Updated dependencies
+  - @badaitech/chaingraph-types@0.1.19
+
 ## 0.1.18
 
 ### Patch Changes