+++ packages/chaingraph-types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @badaitech/chaingraph-types
 
+## 0.1.20
+
+### Patch Changes
+
+- Bump version
+
+## 0.1.19
+
+### Patch Changes
+
+- Bump package
+
 ## 0.1.18
 
 ### Patch Changes